Apply Repository Pattern to recipes with CSV storage
